feat: getCoinUsage api 구현, 페이지 구현
- coin은 데이터 페이지당 12개씩 받으니 쿼리 키 수식 수정 했습니다
